Unit test (#74)
* storage tests

* system info tests

* target tests

* tethering tests

* tracing tests

+++ tot/cdtp/tracing/command.go
@@ -73,7 +73,7 @@ type StartParams struct {
 
 	// Optional. If set, the agent will issue bufferUsage events at this
 	// interval, specified in milliseconds.
-	BufferUsageReportingInterval float64 `json:"bufferUsageReportingInterval,omitempty"`
+	BufferUsageReportingInterval int64 `json:"bufferUsageReportingInterval,omitempty"`
 
 	// Optional. Whether to report trace events as series of dataCollected
 	// events or to save trace to a stream (defaults to `ReportEvents`).
